Relocate BoxQueryEncoding
